# A03_OpenAPI_2206081925.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Marketplace API",
    "description": "Dokumentasi API untuk arsitektur microservices"
  },
  "host": "localhost:8000",
  "basePath": "/api",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/v1/product": {
      "post": {
        "summary": "Tambah produk baru (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name",
                "description",
                "price",
                "quantity_available",
                "category_id"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "quantity_available": {
                  "type": "integer"
                },
                "category_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Produk berhasil ditambahkan"
          }
        }
      }
    },
    "/v2/product/{productId}": {
      "get": {
        "summary": "Dapatkan detail produk berdasarkan ID",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Detail produk ditemukan"
          }
        }
      }
    },
...
  }
}